+++ app/controllers/admin_raw_email_controller.rb
@@ -6,7 +6,7 @@
 # Email: hello@mysociety.org; WWW: http://www.mysociety.org/
 
 class AdminRawEmailController < AdminController
-  before_filter :set_raw_email, only: [:show]
+  before_action :set_raw_email, only: [:show]
 
   def show
     respond_to do |format|