+++ app/controllers/widget_votes_controller.rb
@@ -9,7 +9,7 @@ require 'securerandom'
 
 class WidgetVotesController < ApplicationController
 
-  before_filter :check_widget_config, :find_info_request, :check_prominence
+  before_action :check_widget_config, :find_info_request, :check_prominence
 
   # Track interest in a request from a non-logged in user
   def create