+++ app/controllers/widgets_controller.rb
@@ -7,7 +7,7 @@
 
 class WidgetsController < ApplicationController
 
-  before_filter :check_widget_config, :find_info_request, :check_prominence
+  before_action :check_widget_config, :find_info_request, :check_prominence
 
   def show
     use_secure_headers_override(:allow_frames)