+++ app/models/info_request_batch.rb
@@ -121,6 +121,13 @@ class InfoRequestBatch < ActiveRecord::Base
       mail_message.message_id)
   end
 
+  # Do we consider the InfoRequestBatch to be sent to all authorities?
+  #
+  # Returns a Boolean
+  def sent?
+    sent_at.present?
+  end
+
   # Build an InfoRequest object which is an example of this batch.
   def example_request
     public_body = self.public_bodies.first
@@ -145,6 +152,13 @@ class InfoRequestBatch < ActiveRecord::Base
   # Returns boolean
   def embargo_expiring?
     info_requests.embargo_expiring.any?
+  end
+
+  # Can the Embargo be safely changed?
+  #
+  # Returns a Boolean
+  def can_change_embargo?
+    sent?
   end
 
   # What phases are the requests in this batch in