+++ doc/CHANGES.md
@@ -1,3 +1,18 @@
+# develop
+
+## Highlighted Features
+
+* Prevent censor rules from being unintentionally made permanent when admins
+  edit outgoing messages. Allow admins to see the unredacted outgoing message
+  text on request's admin page and in the associated event log (Liz Conlan)
+
+## Upgrade Notes
+
+* The `hidden_incoming_message` factory has been removed. Use the `:hidden`
+  _trait_ instead if you rely on this in theme specs. See
+  https://github.com/thoughtbot/factory_bot/blob/v4.10.0/GETTING_STARTED.md#traits
+  for more information on traits.
+
 # 0.32.0.0
 
 ## Highlighted Features